Improved on passing pipe file descriptors to child processes.

+++ client.c
@@ -1,12 +1,28 @@
 #include "shared.h"
 
 #include <stdio.h>
+#include <stdlib.h>
 #include <signal.h>
 #include <unistd.h>
 
+static int pipe_fd[2];
+
 int main(int argc, char *argv[])
 {
-    printf("Successfully Started: %s\n", argv[1]);
-    kill(getppid(), SIGCHLD);
+    if(argc != 3) {
+        fprintf(stderr, "Client expects 3 parameters. Only receive %d.\n", argc);
+        exit(EXIT_FAILURE);
+    }
+    sscanf(argv[2], "%d.%d", &pipe_fd[0], &pipe_fd[1]);
+    printf("Successfully Started Station: %s\n", argv[1]);
+    kill(getppid(), SIGUSR1);
+    
+    /*while(1) {
+        printf("Hello from %s\n", argv[1]);
+        fflush(stdout);
+        kill(getppid(), SIGCHLD);
+        sleep(1);
+    }*/
+    
     return 0;
 }
